requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
teemup==1.2.1
pyyaml==6.0.3
